Show contest status in repo README (#482)
This commit improves the visibility of the status of the verification
effort. More specifically, it makes the tools, challenges, and proofs
easier to find  by listing the currently accepted tools and the
challenges, with their reward, status, and link to the proof (once
resolved) in a table in the repo README.

As a personal testimony, once I found the contest blog post, it still
took me a
while before I discovered the challenges. This should improve that
situation.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.